+++ src/main/java/com/ceub/pi/effycityservice/repository/UsuarioGestorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioGestorRepository extends JpaRepository<UsuarioGestor, Integer> {
+public interface UsuarioGestorRepository extends JpaRepository<UsuarioGestor, Long> {
 
 }
